+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/InitializeIntelliJPlatformPluginTask.kt
@@ -17,7 +17,6 @@ import org.jetbrains.intellij.platform.gradle.GradleProperties
 import org.jetbrains.intellij.platform.gradle.get
 import org.jetbrains.intellij.platform.gradle.providers.CurrentPluginVersionValueSource
 import org.jetbrains.intellij.platform.gradle.providers.LatestPluginVersionValueSource
-import org.jetbrains.intellij.platform.gradle.tasks.aware.IntelliJPlatformVersionAware
 import org.jetbrains.intellij.platform.gradle.tasks.aware.ModuleAware
 import org.jetbrains.intellij.platform.gradle.utils.Logger
 import org.jetbrains.intellij.platform.gradle.utils.Version
@@ -38,7 +37,7 @@ private const val CLEAN = "clean"
  * The self-update check can be disabled via [GradleProperties.SelfUpdateCheck] Gradle property.
  */
 @UntrackedTask(because = "Should always run")
-abstract class InitializeIntelliJPlatformPluginTask : DefaultTask(), IntelliJPlatformVersionAware, ModuleAware {
+abstract class InitializeIntelliJPlatformPluginTask : DefaultTask(), ModuleAware {
 
     /**
      * Determines if the operation is running in offline mode and depends on Gradle start parameters.